requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
passlib[bcrypt]
python-jose
mysql-connector-python
joblib
scikit-learn
pandas
numpy